Add TorchScriptable transformer classifier and subword BPE tokenizer (#4566)
* Add torchscript classes for Transformer classifier and subword BPE tokenizer

* Add unit tests

* Remove debugging code

* Resolve comments

* Rerun autoformat.sh

* Update parlai/torchscript/tokenizer.py
